pyvenv-workon should accept "nil" as an argument.
This apparently can happen for certain completing-read replacements,
which return nil instead of "" for the default. In those cases, we
likely should simply use the first environment instead of none,
but as this is a workaround for a bug in those frameworks, I guess this
works for now.

See https://github.com/jorgenschaefer/elpy/issues/144