Override user's shell for hook loading
This is a tricky one. Hooks are sourced into the current shell in
virtualenvwrapper, so they _probably_ should run as the user's
preferred shell. On the other hand, said shell can cause quite
a quoting hell, so using /bin/sh is more likely to keep us sane.

Trying the latter for now. If this breaks for users, we can try and
keep the syntax more compatible by moving the echo line to Python.

Fixes #36