fix python executable used in pyvenv-run-virtualenvwrapper-hook
Hardcoded `python` was used to execute virtualenvwrapper module when
running hooks.

This can fail in pre_activate since the new virtualenv is not yet
activated: there's no guarantee that `python` exists in the system.

Using the configurable variable `pyvenv-virtualenvwrapper-python` is
more correct.

Note that this was happening before commit 74ab422 but that commit
made the problem evident to me since the process output parsing has
changed.